+++ geometry/Building.cpp
@@ -51,15 +51,15 @@
 using namespace std;
 
 
-Building::Building()
-{
-     _caption = "no_caption";
-     _projectFilename = "";
-     _geometryFilename= "";
-     _routingEngine = nullptr;
-     _linkedCellGrid = nullptr;
-     _savePathway = false;
-}
+//Building::Building()
+//{
+//     _caption = "no_caption";
+//     _projectFilename = "";
+//     _geometryFilename= "";
+//     _routingEngine = nullptr;
+//     _linkedCellGrid = nullptr;
+//     _savePathway = false;
+//}
 
 #ifdef _SIMULATOR
 Building::Building(const std::string& filename, const std::string& rootDir, RoutingEngine& engine, PedDistributor& distributor, double linkedCellSize)
@@ -72,7 +72,7 @@ Building::Building(const std::string& filename, const std::string& rootDir, Rout
      //todo: what happens if any of these  methods failed (return false)? throw exception ?
      this->LoadGeometry();
      this->LoadRoutingInfo(filename);
-     this->AddSurroundingRoom();
+     //this->AddSurroundingRoom();
      this->InitGeometry();
      this->LoadTrafficInfo();
      distributor.Distribute(this);
@@ -843,26 +843,37 @@ SubRoom* Building::GetSubRoomByUID( int uid)
      return NULL;
 }
 
-bool Building::IsVisible(Line* l1, Line* l2, bool considerHlines)
-{
+//bool Building::IsVisible(Line* l1, Line* l2, bool considerHlines)
+//{
+//
+//     for(auto&& itr_room: _rooms)
+//     {
+//          for(auto&& itr_subroom: itr_room.second->GetAllSubRooms())
+//          {
+//               if(itr_subroom.second->IsVisible(l1,l2,considerHlines)==false) return false;
+//          }
+//     }
+//     return true;
+//}
 
-     for(auto&& itr_room: _rooms)
+bool Building::IsVisible(const Point& p1, const Point& p2, const std::vector<SubRoom*>& subrooms, bool considerHlines)
+{
+     //loop over all subrooms if none is provided
+     if (subrooms.empty())
      {
-          for(auto&& itr_subroom: itr_room.second->GetAllSubRooms())
+          for(auto&& itr_room: _rooms)
           {
-               if(itr_subroom.second->IsVisible(l1,l2,considerHlines)==false) return false;
+               for(auto&& itr_subroom: itr_room.second->GetAllSubRooms())
+               {
+                    if(itr_subroom.second->IsVisible(p1,p2,considerHlines)==false) return false;
+               }
           }
      }
-     return true;
-}
-
-bool Building::IsVisible(const Point& p1, const Point& p2, bool considerHlines)
-{
-     for(auto&& itr_room: _rooms)
+     else
      {
-          for(auto&& itr_subroom: itr_room.second->GetAllSubRooms())
+          for(auto&& sub: subrooms)
           {
-               if(itr_subroom.second->IsVisible(p1,p2,considerHlines)==false) return false;
+               if(sub and sub->IsVisible(p1,p2,considerHlines)==false) return false;
           }
      }
 
@@ -937,7 +948,6 @@ void Building::InitGrid(double cellSize)
      y_max = y_max + 1*cellSize;
 
      double boundaries[4] = { x_min, x_max, y_min, y_max };
-     int pedsCount = _allPedestians.size();
 
      //no algorithms
      // the domain is made of a single cell
@@ -953,7 +963,8 @@ void Building::InitGrid(double cellSize)
           Log->Write("INFO: \tInitializing the grid with cell size: %f ", cellSize);
      }
 
-     _linkedCellGrid = new LCGrid(boundaries, cellSize, pedsCount);
+     //_linkedCellGrid = new LCGrid(boundaries, cellSize, _allPedestians.size());
+     _linkedCellGrid = new LCGrid(boundaries, cellSize, Pedestrian::GetAgentsCreated());
      _linkedCellGrid->ShallowCopy(_allPedestians);
 
      Log->Write("INFO: \tDone with Initializing the grid ");
@@ -1144,7 +1155,8 @@ void Building::DeletePedestrian(Pedestrian* &ped)
      }
      //update the stats before deleting
      Transition* trans =GetTransitionByUID(ped->GetExitIndex());
-     if(trans) {
+     if(trans)
+     {
           trans->IncreaseDoorUsage(1, ped->GetGlobalTime());
      }
      delete ped;
@@ -1169,9 +1181,12 @@ void Building::AddPedestrian(Pedestrian* ped)
 
 void Building::GetPedestrians(int room, int subroom, std::vector<Pedestrian*>& peds) const
 {
-     for(unsigned int p = 0;p<_allPedestians.size();p++){
-          Pedestrian* ped=_allPedestians[p];
-          if(room==ped->GetRoomID() && subroom==ped->GetSubRoomID())
+     //for(unsigned int p = 0;p<_allPedestians.size();p++){
+     //     Pedestrian* ped=_allPedestians[p];
+
+     for (auto&& ped : _allPedestians)
+     {
+          if ((room == ped->GetRoomID()) && (subroom == ped->GetSubRoomID()))
           {
                peds.push_back(ped);
           }