+++ geometry/SubRoom.cpp
@@ -165,9 +165,29 @@ const vector<int>& SubRoom::GetAllGoalIDs() const
 }
 
 
-void SubRoom::AddWall(const Wall& w)
+bool SubRoom::AddWall(const Wall& w)
 {
+     //check for duplicated walls
+     for(auto&& w1: _walls)
+     {
+          if(w==w1)
+          {
+               Log->Write("ERROR:\t Duplicate wall found in Room/Subroom %d/%d  %s",_roomID,_id ,w.toString().c_str());
+               Log->Write("ERROR:\t will not be added");
+               exit(EXIT_FAILURE);
+               //return false;
+          }
+     }
+     //checking for wall chunks.
+     if(w.Length()<J_TOLERANZ)
+     {
+          Log->Write("ERROR:\t Wall too small (length = %lf) found in Room/Subroom %d/%d  %s",w.Length(),_roomID,_id ,w.toString().c_str());
+          Log->Write("ERROR:\t will not be added");
+          //exit(EXIT_FAILURE);
+          //return false;
+     }
      _walls.push_back(w);
+     return true;
 }
 
 
@@ -197,15 +217,15 @@ void SubRoom::AddTransition(Transition* line)
 
 void SubRoom::AddHline(Hline* line)
 {
-    for(unsigned int i=0;i<_hlines.size();i++){
-        if (line->GetID()==_hlines[i]->GetID()){
-            Log->Write("INFO:\tskipping duplicate hline [%d] with id [%d]",_id,line->GetID());
-            return;
-        }
-    }
+     for(unsigned int i=0;i<_hlines.size();i++){
+          if (line->GetID()==_hlines[i]->GetID()){
+               Log->Write("INFO:\tskipping duplicate hline [%d] with id [%d]",_id,line->GetID());
+               return;
+          }
+     }
 
-    _hlines.push_back(line);
-    _goalIDs.push_back(line->GetUniqueID());
+     _hlines.push_back(line);
+     _goalIDs.push_back(line->GetUniqueID());
 }
 
 const vector<Crossing*>& SubRoom::GetAllCrossings() const
@@ -324,12 +344,13 @@ std::vector<Wall> SubRoom::GetVisibleWalls(const Point & position){
      }
 #if DEBUG
      printf("Leave GetVisiblewalls with %d visible walls\n------------------------------\n",visible_walls.size());
- #endif
+#endif
      return visible_walls;
 }
 
 // like ped_is_visible() but here we can exclude checking intersection
 // with the same wall. This function should check if <position> can see the <Wall>
+//TODO: optimize
 bool SubRoom::IsVisible(const Line &wall, const Point &position)
 {
      // printf("\tEnter wall_is_visible\n");
@@ -359,7 +380,7 @@ bool SubRoom::IsVisible(const Line &wall, const Point &position)
                if(wall_is_vis && ped_wall.IntersectionWith(w)){
                     // fprintf (stdout, "\t\t Wall_is_visible INTERSECTION OBS; L1_P1(%.2f, %.2f), L1_P2(%.2f, %.2f), L2_P1(%.2f, %.2f) L2_P2(%.2f, %.2f)\n", w.GetPoint1().GetX(), w.GetPoint1().GetY(), w.GetPoint2().GetX(), w.GetPoint2().GetY(), ped_wall.GetPoint1().GetX(), ped_wall.GetPoint1().GetY(), ped_wall.GetPoint2().GetX(), ped_wall.GetPoint2().GetY());
                     wall_is_vis = false;
-                    
+
                }
           }
      }
@@ -373,40 +394,43 @@ bool SubRoom::IsVisible(const Line &wall, const Point &position)
 // with the nearest point on the wall IS intersecting with the wall.
 bool SubRoom::IsVisible(const Point& p1, const Point& p2, bool considerHlines)
 {
-     // printf("\t\tEnter ped_is_visible\n");
-     // generate certain connection lines
-     // connecting p1 with p2
-     //Line cl(p1,p2);
-     //Line L2;
-     bool temp =  true;
      //check intersection with Walls
-    for(const auto& wall : _walls) {
-        if (wall.IntersectionWith(p1, p2)) {
-            return false;
-        }
-    }
+     for(const auto& wall : _walls)
+     {
+          if (wall.IntersectionWith(p1, p2))
+          {
+               return false;
+          }
+     }
 
      // printf("\t\t -- ped_is_visible; check obstacles\n");
      //check intersection with obstacles
-    for ( const auto obstacle : _obstacles) {
-        for ( const auto& wall : obstacle->GetAllWalls()) {
-            if (wall.IntersectionWith(p1, p2)) {
-                return false;
-            }
-        }
-    }
-
-
-     // check intersection with other hlines in room
-     if(considerHlines)
-          for(unsigned int i = 0; i < _hlines.size(); i++) {
-               if(_hlines[i]->IsInLineSegment(p1)|| _hlines[i]->IsInLineSegment(p2)) continue;
-               if(temp && (Line*)_hlines[i]->IntersectionWith(p1, p2))
-                    temp = false;
+     for ( const auto& obstacle : _obstacles)
+     {
+          for ( const auto& wall : obstacle->GetAllWalls())
+          {
+               if (wall.IntersectionWith(p1, p2))
+               {
+                    return false;
+               }
           }
+     }
+
+     if(considerHlines)
+     {
+          for(const auto& hline: _hlines)
+          {
+
+               if(hline->IsInLineSegment(p1)|| hline->IsInLineSegment(p2)) continue;
+               if(hline->IntersectionWith(p1, p2))
+               {
+                    return false;
+               }
+          }
+     }
 
      // printf("\t\tLeave ped_is_visible with %d\n",temp);
-     return temp;
+     return true;
 }
 
 bool SubRoom::IsVisible(Line* l1, Line* l2, bool considerHlines)
@@ -455,6 +479,7 @@ bool SubRoom::IsVisible(Line* l1, Line* l2, bool considerHlines)
 }
 
 // this is the case if they share a transition or crossing
+//TODO collect and store the neighbors
 bool SubRoom::IsDirectlyConnectedWith(const SubRoom* sub) const
 {
 
@@ -540,6 +565,103 @@ bool SubRoom::SanityCheck()
                // everything is fine
           }
      }
+     //check if there are overlapping walls
+     for (auto&& w1: _walls)
+     {
+          bool connected=false;
+
+          for (auto&& w2: _walls)
+          {
+               if (w1==w2) continue;
+               if(w1.Overlapp(w2))
+               {
+                    Log->Write("ERROR: Overlapping between walls %s and %s ",w1.toString().c_str(),w2.toString().c_str());
+                    exit(EXIT_FAILURE);
+                    //return false;
+               }
+               connected=connected or w1.ShareCommonPointWith(w2);
+          }
+          //overlapping with lines
+          for(auto&& hline: _hlines)
+          {
+               if(w1.Overlapp(*hline))
+               {
+                    Log->Write("ERROR: Overlapping between wall %s and  Hline %s ",w1.toString().c_str(),hline->toString().c_str());
+                    exit(EXIT_FAILURE);
+                    //return false;
+               }
+          }
+          //overlaping with crossings
+          for(auto&& c: _crossings)
+          {
+               if(w1.Overlapp(*c))
+               {
+                    Log->Write("ERROR: Overlapping between wall %s and  crossing %s ",w1.toString().c_str(),c->toString().c_str());
+                    exit(EXIT_FAILURE);
+                    //return false;
+               }
+               connected=connected or w1.ShareCommonPointWith(*c);
+          }
+          //overlaping with transitions
+          for(auto&& t: _transitions)
+          {
+               if(w1.Overlapp(*t))
+               {
+                    Log->Write("ERROR: Overlapping between wall %s and  transition %s ",w1.toString().c_str(),t->toString().c_str());
+                    exit(EXIT_FAILURE);
+                    //return false;
+               }
+               connected=connected or w1.ShareCommonPointWith(*t);
+          }
+
+          if(not connected)
+          {
+               Log->Write("ERROR: loose wall found %s  in Room/Subroom %d/%d",w1.toString().c_str(),_roomID,_id);
+               //exit(EXIT_FAILURE);
+               //return false;
+          }
+     }
+
+     //check overlaping between hline and crossings
+     for(auto&& h: _hlines)
+     {
+          for(auto&& c: _crossings)
+          {
+               if(h->Overlapp(*c))
+               {
+                    Log->Write("ERROR: Overlapping between hline %s and  crossing %s ",h->toString().c_str(),c->toString().c_str());
+                    exit(EXIT_FAILURE);
+                    //return false;
+               }
+          }
+     }
+     //check overlaping between hlines and transitions
+     for(auto&& h: _hlines)
+     {
+          for(auto&& t: _transitions)
+          {
+               if(h->Overlapp(*t))
+               {
+                    Log->Write("ERROR: Overlapping between hline %s and  transition %s ",h->toString().c_str(),t->toString().c_str());
+                    exit(EXIT_FAILURE);
+                    //return false;
+               }
+          }
+     }
+     //check overlaping between transitions and crossings
+     for(auto&& c: _crossings)
+     {
+          for(auto&& t: _transitions)
+          {
+               if(c->Overlapp(*t))
+               {
+                    Log->Write("ERROR: Overlapping between crossing %s and  transition %s ",c->toString().c_str(),t->toString().c_str());
+                    exit(EXIT_FAILURE);
+                    //return false;
+               }
+          }
+     }
+
      return true;
 }
 
@@ -621,30 +743,55 @@ string NormalSubRoom::WriteSubRoom() const
           char wall[CLENGTH] = "";
           geometry.append("\t\t<wall>\n");
           sprintf(wall, "\t\t\t<point xPos=\"%.2f\" yPos=\"%.2f\" zPos=\"%.2f\"/>\n",
-                  (w.GetPoint1().GetX()) * FAKTOR,
-                  (w.GetPoint1().GetY()) * FAKTOR,
-                  GetElevation(w.GetPoint1())*FAKTOR);
+                    (w.GetPoint1().GetX()) * FAKTOR,
+                    (w.GetPoint1().GetY()) * FAKTOR,
+                    GetElevation(w.GetPoint1())*FAKTOR);
           geometry.append(wall);
           sprintf(wall, "\t\t\t<point xPos=\"%.2f\" yPos=\"%.2f\" zPos=\"%.2f\"/>\n",
-                  (w.GetPoint2().GetX()) * FAKTOR,
-                  (w.GetPoint2().GetY()) * FAKTOR,
-                  GetElevation(w.GetPoint2())*FAKTOR);
+                    (w.GetPoint2().GetX()) * FAKTOR,
+                    (w.GetPoint2().GetY()) * FAKTOR,
+                    GetElevation(w.GetPoint2())*FAKTOR);
           geometry.append(wall);
           geometry.append("\t\t</wall>\n");
 
           s.append(geometry);
           //s.append(GetWall(j).Write());
      }
+
      //add the subroom caption
-     Point pos = GetCentroid();
+     const Point& pos = GetCentroid();
      char tmp[CLENGTH];
-     sprintf(tmp, "\t\t<label centerX=\"%.2f\" centerY=\"%.2f\" centerZ=\"0\" text=\"%d\" color=\"100\" />\n"
-             , pos.GetX() * FAKTOR, pos.GetY() * FAKTOR, GetSubRoomID());
+     sprintf(tmp, "\t\t<label centerX=\"%.2f\" centerY=\"%.2f\" centerZ=\"%.2f\" text=\"%d\" color=\"100\" />\n"
+               , pos.GetX() * FAKTOR, pos.GetY() * FAKTOR, GetElevation(pos)*FAKTOR,GetSubRoomID());
      s.append(tmp);
 
      //write the obstacles
-     for( unsigned int j=0; j<GetAllObstacles().size(); j++) {
-          s.append(GetAllObstacles()[j]->Write());
+     for (auto&& obst : GetAllObstacles())
+     {
+          for (auto&& w: obst->GetAllWalls())
+          {
+               char wall[CLENGTH] = "";
+               s.append("\t\t<wall>\n");
+               sprintf(wall, "\t\t\t<point xPos=\"%.2f\" yPos=\"%.2f\" zPos=\"%.2f\"/>\n",
+                         (w.GetPoint1().GetX()) * FAKTOR,
+                         (w.GetPoint1().GetY()) * FAKTOR,
+                         GetElevation(w.GetPoint1())*FAKTOR);
+               s.append(wall);
+               sprintf(wall, "\t\t\t<point xPos=\"%.2f\" yPos=\"%.2f\" zPos=\"%.2f\"/>\n",
+                         (w.GetPoint2().GetX()) * FAKTOR,
+                         (w.GetPoint2().GetY()) * FAKTOR,
+                         GetElevation(w.GetPoint2())*FAKTOR);
+               s.append(wall);
+               s.append("\t\t</wall>\n");
+          }
+
+          const Point& pos = obst->GetCentroid();
+
+          //add the obstacle caption
+          char tmp[CLENGTH];
+          sprintf(tmp, "\t\t<label centerX=\"%.2f\" centerY=\"%.2f\" centerZ=\"%.2f\" text=\"%d\" color=\"100\" />\n"
+                    , pos.GetX() * FAKTOR, pos.GetY() * FAKTOR,GetElevation(pos)*FAKTOR ,obst->GetId());
+          s.append(tmp);
      }
 
      return s;
@@ -717,8 +864,8 @@ bool NormalSubRoom::ConvertLineToPoly(vector<Line*> goals)
      if ((tmpPoly[0] - point).Norm() > J_TOLERANZ) {
           char tmp[CLENGTH];
           sprintf(tmp, "ERROR: \tNormalSubRoom::ConvertLineToPoly(): SubRoom %d Room %d Anfangspunkt ungleich Endpunkt!!!\n"
-                  "\t(%f, %f) != (%f, %f)\n", GetSubRoomID(), GetRoomID(), tmpPoly[0].GetX(), tmpPoly[0].GetY(), point.GetX(),
-                  point.GetY());
+                    "\t(%f, %f) != (%f, %f)\n", GetSubRoomID(), GetRoomID(), tmpPoly[0].GetX(), tmpPoly[0].GetY(), point.GetX(),
+                    point.GetY());
           Log->Write(tmp);
           sprintf(tmp, "ERROR: \tDistance between the points: %lf !!!\n", (tmpPoly[0] - point).Norm());
           Log->Write(tmp);
@@ -736,7 +883,7 @@ bool NormalSubRoom::ConvertLineToPoly(vector<Line*> goals)
 int NormalSubRoom::WhichQuad(const Point& vertex, const Point& hitPos) const
 {
      return (vertex.GetX() > hitPos.GetX()) ? ((vertex.GetY() > hitPos.GetY()) ? 1 : 4) :
-                 ((vertex.GetY() > hitPos.GetY()) ? 2 : 3);
+               ((vertex.GetY() > hitPos.GetY()) ? 2 : 3);
 
 }
 
@@ -745,7 +892,7 @@ int NormalSubRoom::WhichQuad(const Point& vertex, const Point& hitPos) const
 double NormalSubRoom::Xintercept(const Point& point1, const Point& point2, double hitY) const
 {
      return (point2.GetX() - (((point2.GetY() - hitY) * (point1.GetX() - point2.GetX())) /
-                              (point1.GetY() - point2.GetY())));
+               (point1.GetY() - point2.GetY())));
 }
 
 
@@ -783,7 +930,7 @@ bool NormalSubRoom::IsInSubRoom(const Point& ped) const
           case -3: // MOVING BACK 3 IS LIKE MOVING FORWARD 1
                delta = 1;
                break;
-               default:break;
+          default:break;
           }
           /* ADD IN THE DELTA */
           total += delta;
@@ -848,14 +995,14 @@ string Stair::WriteSubRoom() const
           char wall[CLENGTH] = "";
           geometry.append("\t\t<wall>\n");
           sprintf(wall, "\t\t\t<point xPos=\"%.2f\" yPos=\"%.2f\" zPos=\"%.2f\"/>\n",
-                  (w.GetPoint1().GetX()) * FAKTOR,
-                  (w.GetPoint1().GetY()) * FAKTOR,
-                  GetElevation(w.GetPoint1())*FAKTOR);
+                    (w.GetPoint1().GetX()) * FAKTOR,
+                    (w.GetPoint1().GetY()) * FAKTOR,
+                    GetElevation(w.GetPoint1())*FAKTOR);
           geometry.append(wall);
           sprintf(wall, "\t\t\t<point xPos=\"%.2f\" yPos=\"%.2f\" zPos=\"%.2f\"/>\n",
-                  (w.GetPoint2().GetX()) * FAKTOR,
-                  (w.GetPoint2().GetY()) * FAKTOR,
-                  GetElevation(w.GetPoint2())*FAKTOR);
+                    (w.GetPoint2().GetX()) * FAKTOR,
+                    (w.GetPoint2().GetY()) * FAKTOR,
+                    GetElevation(w.GetPoint2())*FAKTOR);
           geometry.append(wall);
           geometry.append("\t\t</wall>\n");
 
@@ -867,12 +1014,12 @@ string Stair::WriteSubRoom() const
      Point pos = GetCentroid();
      char tmp_c[CLENGTH];
      sprintf(tmp_c, "\t\t<sphere centerX=\"%.2f\" centerY=\"%.2f\" centerZ=\"%.2f\" radius=\"%.2f\" color=\"100\" />\n"
-             , GetUp().GetX() * FAKTOR, GetUp().GetY() * FAKTOR,GetElevation(GetUp())*FAKTOR, 0.2*FAKTOR);
+               , GetUp().GetX() * FAKTOR, GetUp().GetY() * FAKTOR,GetElevation(GetUp())*FAKTOR, 0.2*FAKTOR);
      s.append(tmp_c);
 
      //add the subroom caption
      sprintf(tmp_c, "\t\t<label centerX=\"%.2f\" centerY=\"%.2f\" centerZ=\"%.2f\" text=\"%d\" color=\"100\" />\n"
-             , pos.GetX() * FAKTOR, pos.GetY() * FAKTOR,GetElevation(pos)*FAKTOR ,GetSubRoomID());
+               , pos.GetX() * FAKTOR, pos.GetY() * FAKTOR,GetElevation(pos)*FAKTOR ,GetSubRoomID());
      s.append(tmp_c);
 
      return s;
@@ -981,7 +1128,7 @@ bool Stair::ConvertLineToPoly(vector<Line*> goals)
           x2 = aktPoint->GetX();
           y2 = aktPoint->GetY();
           sprintf(tmp, "ERROR: \tStair::ConvertLineToPoly(): SubRoom %d Room %d Anfangspunkt ungleich Endpunkt!!!\n"
-                  "\t(%f, %f) != (%f, %f)\n", GetSubRoomID(), GetRoomID(), x1, y1, x2, y2);
+                    "\t(%f, %f) != (%f, %f)\n", GetSubRoomID(), GetRoomID(), x1, y1, x2, y2);
           Log->Write(tmp);
           return false;
      }
@@ -989,7 +1136,7 @@ bool Stair::ConvertLineToPoly(vector<Line*> goals)
      if (orgPoly.size() != 4) {
           char tmp[CLENGTH];
           sprintf(tmp, "ERROR: \tStair::ConvertLineToPoly(): Stair %d Room %d ist kein Viereck!!!\n"
-                  "Anzahl Ecken: %d\n", GetSubRoomID(), (int)GetRoomID(), (int)orgPoly.size());
+                    "Anzahl Ecken: %d\n", GetSubRoomID(), (int)GetRoomID(), (int)orgPoly.size());
           Log->Write(tmp);
           return false;
      }